Rename schemas directory to 'schema' and update validate_config_data to support JSON output
