+++ src/illuminator/schema/simulation.example.yaml
