+++ src/illuminator/schema/validate_yaml.py
@@ -1,10 +1,10 @@
 from ruamel.yaml import YAML
 import sys
 import json
-from illuminator.schemas.simulation import schema
+from illuminator.schema.simulation import schema
 
 # this if a yaml file complies with the modelling schema
-_file = open('simple_test.yaml', 'r')
+_file = open('simple_test2.yaml', 'r')
 
 yaml = YAML(typ='safe')
 data = yaml.load(_file)
@@ -13,5 +13,6 @@ val = schema.validate(data)
 print(val)
 
 json_data = json.dumps(data, indent=4)
-print(json_data)
+
+print(type(json_data))
